💄 fix css styling and simplify landingpage to have banner image, change links to signup page and landing page
