+++ client/src/App.css
@@ -13,7 +13,7 @@
   --primary-grey: #212735;
   --secondary-grey: #d9d9d9;
   --secondary-grey-light: #eeeeee;
-  --third-grey: #2e457c95;
+  --third-grey: #2e457ced;
   --button-light: #a8b2ff;
   --button-dark: #3e435f;
   --third-purple: #c8b6ff;
@@ -28,14 +28,12 @@
   font-family: "Nunito", sans-serif;
 }
 
-body,
 html {
   height: 100%;
   background: linear-gradient(
-    135deg,
+    150deg,
     var(--background-black) 0%,
     var(--third-grey) 24%,
     var(--primary-grey) 100%
   );
-  background-repeat: no-repeat;
 }
