+++ client/src/pages/SignUp.jsx
@@ -1,15 +1,18 @@
+import { Link } from "react-router-dom";
 import UserForm from "../components/UserForm";
 
 function SignUp() {
-  return(
-    <div>
-  <div className="sign-up-header">
-    <button type="button" className="go-back-button">Go back</button>
-    <h1 className="sign-up-text">Sign Up !</h1>
-    </div>
-    <UserForm />
-    </div>
-  )
+  return (
+    <section className="sign-up-page">
+      <div className="sign-up-header">
+        <Link to="/landing" className="go-back-button">
+          Go back
+        </Link>
+        <h1 className="sign-up-text">Sign Up !</h1>
+      </div>
+      <UserForm />
+    </section>
+  );
 }
 
 export default SignUp;
